Moved config to dedicated package
